fix: int para long long no 1176_fibonacci_em_vetor

+++ beecrowd/revisao/1176_fibonacci_em_vetor.cpp
@@ -25,7 +25,7 @@ using namespace std;
 int main()
 {
     //declaração de variáveis
-    int fibonacci[61];
+    long long fibonacci[61];
     int t, n;
 
     //entrada de dados
